+++ messages/en/core.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'SAVE' => 'Save',
+    'SAVE_SUCCESS' => 'Saving has been completed successfully!',
+    'SETTINGS' => 'Settings',
+    'PARAMETER_NOT_FOUND' => 'Parameter "{module}.{name}" not found.',
+    'INVALID_PARAMETER_VALUE' => 'Invalid value for "{module}.{name}" parameter.'
+];