+++ messages/ru/core.php
@@ -5,4 +5,5 @@ return [
     'SAVE_SUCCESS' => 'Сохранение выполненно успешно!',
     'SETTINGS' => 'Настройки',
     'PARAMETER_NOT_FOUND' => 'Параметр "{module}.{name}" не найден.',
+    'INVALID_PARAMETER_VALUE' => 'Неверное значение параметра "{module}.{name}".'
 ];